remove: App.tsxを削除し、src/App.tsxに移動しました。package.jsonとpackage-lock.jsonを更新しました。
